requirements.txt
# Package Dependencies
numpy
pandas
scikit-learn
scipy
h5py
numpy
neurodsp
decorator
praatio
PyYAML

# Optional
# mne

# Random Dependencies that maybe removed
ipywidgets


